requirements.txt
django-disqus
